cli: implement more flags in the list command
Introduces all the flags functionality, except format, sort and json. Some other features, such as the session headers, were left with a todo mark.

+++ client/operations/get_files_responses.go
@@ -385,7 +385,7 @@ type GetFilesOKBodyItemsItems0Size struct {
 	HumanReadable string `json:"human_readable,omitempty"`
 
 	// raw
-	Raw float64 `json:"raw,omitempty"`
+	Raw int64 `json:"raw,omitempty"`
 }
 
 // Validate validates this get files o k body items items0 size
